Retire use of ra_file_handle
As it adds fixed overhead to each disk IO operation and duplicates
information that is better accessed by the operation system.

To compensate this adds a "bytes written" counter to the wal and
segment writer processes respectively. This will allow of inspection
of Ra component specific IO use to some degree.

Also refactored the WAL sync method to also include an fadvise
instruction to purge page cache after fsync as we do no need it.